Add optimized implementation of AES for Apple platform
